add dengine meta library
this lib is pretty much the meta package which include all stuff. it
provides some very basic high level functions which will be used for a
quick firing up of a window and rendering text.

It also allows merging other libs into it for sanity and keeping
directory usage to a minimum. Note that merging requires -fPIC