add renderbuffer attach and fb readback
note that attaching framebuffer does not attempt to change the
framebuffer client side texture arrays as with a texture.

THIS FUNCTIONALITY WILL BE REMOVED IN A FUTURE BUILD!

It causes a myriad of issues like stackoveruse (i invented that 😆) its
much saner to just readback to a texture.
its also not very compatible with ES 2.0/3.0 where its illegal to attach
textures to framebuffers...