remove scene_glarea refs
we essentialy just need 1 GtkGLArea for rendering then switch its mode
from scene view to game view

editors like Unity and Unreal have a way to show both the scene and game
view. a very cool feature on paper but very tricky to implement with
Gtk. you can use shared contexts which are yet another black hole you
really don't want to deal with 😡

You could however readback the glarea  to some memory and upload it to a
shared context, readback it to another GL area and boom. this is very
slow but that's my current hypothesis as of now