0.5.0 : Release Notes
This release merges feature-script-python to main and introduces python
scripting to upstream 🍻

Please see some example scripts in assets/scripts. See the main
README.md on how to get scripting to work with VSCode. Its seemless on
Linux but tricky on Windows

Other changes include:
**DENGITOR**:
- use callback for receiving logging. Any messages that go thru `printf`
or `print`(Python) will be redirected to Log. This also has the same
effect with android

**TESTS**:
- add android ecs test (pure ecs. no C code is moving anything!)
- add python no gui test

**LIB**:
- dengine-utils/logging:
	- added a logging thread which redirects stdout and stderr to callbacks. Useful for `printf` to android logcat or printf to dengitor. On android its kinda rough as it only prints once the app exits
- dengine-utils/filesys:
	- reorder how filesys detects assets dir. Uses srcdir as the last resort
- dengine-model/model:
	- can load a model from memory

**ASSETS**:
- downscaled sea cubemap to 512x512 and sunset.hdr 1024x512 to reduce
disk usage