android: recover when app is resumed
this mainly involved recreating EGLSurface when app switched or when
home is pressed.

android-util was also reworked to fire callbacks so we can catch app
pause

android_main reworked too to freeze main but still poll events so we can
catch the app resuming. waitevents might be used here but it not working
well currently

MULTITHREADED EVENT LOOP IS STILL BROKEN AND STILL RUNNING ON MAIN
THREAD!!!

TODO: when back key is pressed, then app is launched, it fails to init